It fixes the bug #1198858
properties:  pathAccessedDate, pathModifiedDate, pathCreatedDate  and pathIsWritable which do not have a "change" signal will be removed.
They will be replaced by Q_INVOKABLE functions:
                           curPathAccessedDate(), curPathModifiedDate(), curPathCreatedDate() and curPathIsWritable()

This change is temporary and allows it to work with current properties and with the new Q_INVOKABLE functions.
Once the plugin with new functions is released, it can be changed again to remove references to these properties.

New functions can be tested with lp:~carlos-mazieri/ubuntu-filemanager-app/model-bug-properties-Unknown

If this change will be accepted, after it be released a merge proposal from the branch above to the plugin branch will be requested.